Completed VkQueue implementation

+++ Magnesium.Vulkan/Structs/VkPresentInfoKHR.cs
@@ -10,11 +10,10 @@ namespace Magnesium.Vulkan
 		public VkStructureType sType { get; set; }
 		public IntPtr pNext { get; set; }
 		public UInt32 waitSemaphoreCount { get; set; }
-		public UInt64 pWaitSemaphores { get; set; }
+		public IntPtr pWaitSemaphores { get; set; } // UInt64[]
 		public UInt32 swapchainCount { get; set; }
 		public IntPtr pSwapchains { get; set; } // array
-		public UInt32 pImageIndices { get; set; } 
-		// Reused for results
+		public IntPtr pImageIndices { get; set; } // UInt32[]
 		public IntPtr pResults { get; set; } // array
 	}
 }
